Trim whitespace from properties and values
Incidentally, this provides a work-around to the inability to use the
same property name twice in a selector body (because JavaScript
requires Object keys to be unique). You can just add whitespace
characters to the beginning or end of a property to act as a unique
key; the trim in the output will remove that extra whitespace anyway.